failure conditions, uptime ticker
Added flexibility for manual fan speed control. The scraper would fail, due to how webman handles displaying fan speed percentages.
Instead of displaying all zeroes on a failed scrape, the wallpaper will keep the results of the last scrape until the next successful scrape.
The uptime now has a 'ticker', it will tick along and keep in sync with the PS3, as best as it can.